similarly split FieldMapperTest into FieldMapperMockTest
