+++ core/src/test/java/org/jdbi/v3/core/mapper/reflect/FieldMapperTest.java
@@ -13,207 +13,34 @@
  */
 package org.jdbi.v3.core.mapper.reflect;
 
-import java.math.BigDecimal;
-import java.sql.ResultSet;
-import java.sql.ResultSetMetaData;
 import java.sql.SQLException;
 
 import javax.annotation.Nullable;
 
 import org.jdbi.v3.core.Handle;
-import org.jdbi.v3.core.HandleAccess;
 import org.jdbi.v3.core.SampleBean;
-import org.jdbi.v3.core.ValueType;
 import org.jdbi.v3.core.mapper.Nested;
 import org.jdbi.v3.core.mapper.PropagateNull;
 import org.jdbi.v3.core.mapper.RowMapper;
-import org.jdbi.v3.core.mapper.ValueTypeMapper;
 import org.jdbi.v3.core.rule.H2DatabaseRule;
-import org.jdbi.v3.core.statement.StatementContext;
-import org.jdbi.v3.core.statement.StatementContextAccess;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
-import org.mockito.Mock;
-import org.mockito.junit.MockitoJUnit;
-import org.mockito.junit.MockitoRule;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.Mockito.when;
 
 public class FieldMapperTest {
-
-    @Rule
-    public MockitoRule mockitoRule = MockitoJUnit.rule();
-
     @Rule
     public H2DatabaseRule dbRule = new H2DatabaseRule().withSomething();
 
-    @Mock
-    ResultSet resultSet;
-
-    @Mock
-    ResultSetMetaData resultSetMetaData;
-
-    Handle handle = HandleAccess.createHandle();
-    StatementContext ctx = StatementContextAccess.createContext(handle);
+    Handle handle;
 
     RowMapper<SampleBean> mapper = FieldMapper.of(SampleBean.class);
 
     @Before
-    public void setUpMocks() throws SQLException {
-        when(resultSet.getMetaData()).thenReturn(resultSetMetaData);
-    }
-
-    private void mockColumns(String... columns) throws SQLException {
-        when(resultSetMetaData.getColumnCount()).thenReturn(columns.length);
-        for (int i = 0; i < columns.length; i++) {
-            when(resultSetMetaData.getColumnLabel(i + 1)).thenReturn(columns[i]);
-        }
-    }
-
-    @Test
-    public void shouldSetValueOnPrivateField() throws Exception {
-        mockColumns("longField");
-
-        Long aLongVal = 100L;
-        when(resultSet.getLong(1)).thenReturn(aLongVal);
-        when(resultSet.wasNull()).thenReturn(false);
-
-        SampleBean sampleBean = mapper.map(resultSet, ctx);
-
-        assertThat(sampleBean.getLongField()).isEqualTo(aLongVal);
-    }
-
-    @Test
-    public void shouldHandleEmptyResult() throws Exception {
-        mockColumns();
-
-        SampleBean sampleBean = mapper.map(resultSet, ctx);
-
-        assertThat(sampleBean).isNotNull();
-    }
-
-    @Test
-    public void shouldBeCaseInSensitiveOfColumnAndFieldNames() throws Exception {
-        mockColumns("LoNgfielD");
-
-        Long aLongVal = 100L;
-        when(resultSet.getLong(1)).thenReturn(aLongVal);
-        when(resultSet.wasNull()).thenReturn(false);
-
-        SampleBean sampleBean = mapper.map(resultSet, ctx);
-
-        assertThat(sampleBean.getLongField()).isEqualTo(aLongVal);
-
-    }
-
-    @Test
-    public void shouldHandleNullValue() throws Exception {
-        mockColumns("LoNgfielD");
-
-        when(resultSet.getLong(1)).thenReturn(0L);
-        when(resultSet.wasNull()).thenReturn(true);
-
-        SampleBean sampleBean = mapper.map(resultSet, ctx);
-
-        assertThat(sampleBean.getLongField()).isNull();
-    }
-
-    @Test
-    public void shouldThrowOnTotalMismatch() throws Exception {
-        mockColumns("somethingElseEntirely");
-        assertThatThrownBy(() -> mapper.map(resultSet, ctx)).isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
-    public void shouldSetValuesOnAllFieldAccessTypes() throws Exception {
-        mockColumns("longField", "protectedStringField", "packagePrivateIntField", "privateBigDecimalField");
-
-        Long aLongVal = 100L;
-        String aStringVal = "something";
-        int aIntVal = 1;
-        BigDecimal aBigDecimal = BigDecimal.TEN;
-        when(resultSet.getLong(1)).thenReturn(aLongVal);
-        when(resultSet.getString(2)).thenReturn(aStringVal);
-        when(resultSet.getInt(3)).thenReturn(aIntVal);
-        when(resultSet.getBigDecimal(4)).thenReturn(aBigDecimal);
-        when(resultSet.wasNull()).thenReturn(false);
-
-        SampleBean sampleBean = mapper.map(resultSet, ctx);
-
-        assertThat(sampleBean.getLongField()).isEqualTo(aLongVal);
-        assertThat(sampleBean.getPrivateBigDecimalField()).isEqualTo(aBigDecimal);
-        assertThat(sampleBean.getPackagePrivateIntField()).isEqualTo(aIntVal);
-        assertThat(sampleBean.getProtectedStringField()).isEqualTo(aStringVal);
-    }
-
-    @Test
-    public void shouldSetValuesInSuperClassFields() throws Exception {
-        mockColumns("longField", "blongField");
-
-        Long aLongVal = 100L;
-        Long bLongVal = 200L;
-
-        when(resultSet.getLong(1)).thenReturn(aLongVal);
-        when(resultSet.getLong(2)).thenReturn(bLongVal);
-        when(resultSet.wasNull()).thenReturn(false);
-
-        RowMapper<DerivedBean> mapper = FieldMapper.of(DerivedBean.class);
-
-        DerivedBean derivedBean = mapper.map(resultSet, ctx);
-
-        assertThat(derivedBean.getLongField()).isEqualTo(aLongVal);
-        assertThat(derivedBean.getBlongField()).isEqualTo(bLongVal);
-    }
-
-    @Test
-    public void shouldUseRegisteredMapperForUnknownPropertyType() throws Exception {
-        handle.registerColumnMapper(new ValueTypeMapper());
-
-        mockColumns("longField", "valueTypeField");
-
-        when(resultSet.getLong(1)).thenReturn(123L);
-        when(resultSet.getString(2)).thenReturn("foo");
-        when(resultSet.wasNull()).thenReturn(false);
-
-        SampleBean sampleBean = mapper.map(resultSet, ctx);
-
-        Long expected = 123L;
-        assertThat(sampleBean.getLongField()).isEqualTo(expected);
-        assertThat(sampleBean.getValueTypeField()).isEqualTo(ValueType.valueOf("foo"));
-    }
-
-    @Test
-    public void shouldThrowOnPropertyTypeWithoutRegisteredMapper() throws Exception {
-        mockColumns("longField", "valueTypeField");
-
-        when(resultSet.getLong(1)).thenReturn(123L);
-        when(resultSet.getObject(2)).thenReturn(new Object());
-        when(resultSet.wasNull()).thenReturn(false);
-
-        assertThatThrownBy(() -> mapper.map(resultSet, ctx)).isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
-    public void shouldNotThrowOnMismatchedColumns() throws Exception {
-        mockColumns("longField", "extraColumn");
-
-        Long expected = 666L;
-        when(resultSet.getLong(1)).thenReturn(expected);
-        when(resultSet.getString(2)).thenReturn("foo");
-
-        SampleBean sampleBean = mapper.map(resultSet, ctx);
-
-        assertThat(sampleBean.getLongField()).isEqualTo(expected);
-    }
-
-    @Test
-    public void shouldThrowOnMismatchedColumnsStrictMatch() throws Exception {
-        ctx.getConfig(ReflectionMappers.class).setStrictMatching(true);
-        mockColumns("longField", "misspelledField");
-        assertThatThrownBy(() -> mapper.map(resultSet, ctx)).isInstanceOf(IllegalArgumentException.class);
+    public void openHandle() throws SQLException {
+        handle = dbRule.getSharedHandle();
     }
 
     static class ColumnNameThing {
@@ -226,7 +53,6 @@ public class FieldMapperTest {
 
     @Test
     public void testColumnNameAnnotation() {
-        Handle handle = dbRule.getSharedHandle();
         handle.execute("insert into something (id, name) values (1, 'foo')");
 
         ColumnNameThing thing = handle.createQuery("select * from something")
@@ -239,7 +65,6 @@ public class FieldMapperTest {
 
     @Test
     public void testNested() {
-        Handle handle = dbRule.getSharedHandle();
         handle.execute("insert into something (id, name) values (1, 'foo')");
 
         assertThat(handle
@@ -253,7 +78,6 @@ public class FieldMapperTest {
 
     @Test
     public void testNestedStrict() {
-        Handle handle = dbRule.getSharedHandle();
         handle.getConfig(ReflectionMappers.class).setStrictMatching(true);
         handle.registerRowMapper(FieldMapper.factory(NestedThing.class));
 
@@ -282,7 +106,6 @@ public class FieldMapperTest {
 
     @Test
     public void testNestedPresent() {
-        Handle handle = dbRule.getSharedHandle();
         assertThat(handle
             .registerRowMapper(FieldMapper.factory(NullableNestedThing.class))
             .select("SELECT 42 as testValue, 2 as i, '3' as s")
@@ -294,7 +117,6 @@ public class FieldMapperTest {
 
     @Test
     public void testNestedHalfPresent() {
-        Handle handle = dbRule.getSharedHandle();
         assertThat(handle
             .registerRowMapper(FieldMapper.factory(NullableNestedThing.class))
             .select("SELECT 42 as testValue, '3' as s")
@@ -306,7 +128,6 @@ public class FieldMapperTest {
 
     @Test
     public void testNestedAbsent() {
-        Handle handle = dbRule.getSharedHandle();
         assertThat(handle
             .registerRowMapper(FieldMapper.factory(NullableNestedThing.class))
             .select("SELECT 42 as testValue")
@@ -318,7 +139,6 @@ public class FieldMapperTest {
 
     @Test
     public void testNullableColumnAbsentButNestedPresent() {
-        Handle handle = dbRule.getSharedHandle();
         assertThat(handle
             .registerRowMapper(FieldMapper.factory(NullableNestedThing.class))
             .select("SELECT 's' s, 1 i")
@@ -330,7 +150,6 @@ public class FieldMapperTest {
 
     @Test
     public void testNoRecognizedColumns() {
-        Handle handle = dbRule.getSharedHandle();
         assertThatThrownBy(() -> handle
             .registerRowMapper(FieldMapper.factory(NullableNestedThing.class))
             .select("SELECT 'foo' bar")
@@ -358,7 +177,6 @@ public class FieldMapperTest {
 
     @Test
     public void testNestedPrefix() {
-        Handle handle = dbRule.getSharedHandle();
         handle.execute("insert into something (id, name) values (1, 'foo')");
 
         assertThat(handle
@@ -372,7 +190,6 @@ public class FieldMapperTest {
 
     @Test
     public void testNestedPrefixStrict() {
-        Handle handle = dbRule.getSharedHandle();
         handle.getConfig(ReflectionMappers.class).setStrictMatching(true);
         handle.registerRowMapper(FieldMapper.factory(NestedPrefixThing.class));
 
@@ -408,8 +225,7 @@ public class FieldMapperTest {
     }
 
     @Test
-    public void testPropagateNull() {
-        Handle handle = dbRule.getSharedHandle();
+    public void propagateNull() {
         assertThat(handle
             .registerRowMapper(FieldMapper.factory(PropagateNullThing.class))
             .select("SELECT null as testValue, 'foo' as s")
@@ -419,8 +235,7 @@ public class FieldMapperTest {
     }
 
     @Test
-    public void testPropagateNotNull() {
-        Handle handle = dbRule.getSharedHandle();
+    public void propagateNotNull() {
         assertThat(handle
             .registerRowMapper(FieldMapper.factory(PropagateNullThing.class))
             .select("SELECT 42 as testValue, 'foo' as s")
@@ -431,8 +246,7 @@ public class FieldMapperTest {
     }
 
     @Test
-    public void testNestedPropagateNull() {
-        Handle handle = dbRule.getSharedHandle();
+    public void nestedPropagateNull() {
         assertThat(handle
             .registerRowMapper(FieldMapper.factory(NestedPropagateNullThing.class))
             .select("SELECT 42 as integerValue, null as testValue, 'foo' as s")
@@ -443,8 +257,7 @@ public class FieldMapperTest {
     }
 
     @Test
-    public void testNestedPropagateNotNull() {
-        Handle handle = dbRule.getSharedHandle();
+    public void nestedPropagateNotNull() {
         assertThat(handle
             .registerRowMapper(FieldMapper.factory(NestedPropagateNullThing.class))
             .select("SELECT 42 as integerValue, 60 as testValue, 'foo' as s")
